requirements.txt
streamlit
pillow
google-generativeai
pandas
